+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
-All notable changes to the "amp-jump" extension will be documented in this file.
+## v0.1.0 (??)
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- Better label position lookup
+    - Added ignore symbols: (" ", ",", ".", "(", ")", "\"", "'", "´", "\t", "-", "/", ";")
+    - Words after an ignore symbol will have an label
 
-## [Unreleased]
+## v0.0.1 (2026-01-16)
 
 - Initial release